Make onSubmit a prop
